move spark environment test
